chtml/html/testdata/webkit/tests16.dat

#data
<!doctype html><script>
#errors
(1,23): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>

#data
<!doctype html><script>a
#errors
(1,24): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "a"

#data
<!doctype html><script><
#errors
(1,24): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "<"

#data
<!doctype html><script></
#errors
(1,25): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "</"

#data
<!doctype html><script></S
#errors
(1,26): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "</S"

#data
<!doctype html><script></SC
#errors
(1,27): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "</SC"

#data
<!doctype html><script></SCR
#errors
(1,28): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "</SCR"

#data
<!doctype html><script></SCRI
#errors
(1,29): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "</SCRI"

#data
<!doctype html><script></SCRIP
#errors
(1,30): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "</SCRIP"

#data
<!doctype html><script></SCRIPT
#errors
(1,31): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "</SCRIPT"

#data
<!doctype html><script></SCRIPT 
#errors
(1,32): expected-attribute-name-but-got-eof
(1,32): expected-named-closing-tag-but-got-eof
#new-errors
(1:33) eof-in-tag
#document
| <!DOCTYPE html>
| <script>

#data
<!doctype html><script></s
#errors
(1,26): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "</s"

#data
<!doctype html><script></sc
#errors
(1,27): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "</sc"

#data
<!doctype html><script></scr
#errors
(1,28): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "</scr"

#data
<!doctype html><script></scri
#errors
(1,29): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "</scri"

#data
<!doctype html><script></scrip
#errors
(1,30): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "</scrip"

#data
<!doctype html><script></script
#errors
(1,31): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "</script"

#data
<!doctype html><script></script 
#errors
(1,32): expected-attribute-name-but-got-eof
(1,32): expected-named-closing-tag-but-got-eof
#new-errors
(1:33) eof-in-tag
#document
| <!DOCTYPE html>
| <script>

#data
<!doctype html><script><!
#errors
(1,25): expected-script-data-but-got-eof
(1,25): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "<!"

#data
<!doctype html><script><!a
#errors
(1,26): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "<!a"

#data
<!doctype html><script><!-
#errors
(1,26): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "<!-"

#data
<!doctype html><script><!-a
#errors
(1,27): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "<!-a"

#data
<!doctype html><script><!--
#errors
(1,27): expected-named-closing-tag-but-got-eof
(1,27): unexpected-eof-in-text-mode
#new-errors
(1:28) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--"

#data
<!doctype html><script><!--a
#errors
(1,28): expected-named-closing-tag-but-got-eof
(1,28): unexpected-eof-in-text-mode
#new-errors
(1:29) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--a"

#data
<!doctype html><script><!--<
#errors
(1,28): expected-named-closing-tag-but-got-eof
(1,28): unexpected-eof-in-text-mode
#new-errors
(1:29) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<"

#data
<!doctype html><script><!--<a
#errors
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<a"

#data
<!doctype html><script><!--</
#errors
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--</"

#data
<!doctype html><script><!--</script
#errors
(1,35): expected-named-closing-tag-but-got-eof
(1,35): unexpected-eof-in-text-mode
#new-errors
(1:36) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--</script"

#data
<!doctype html><script><!--</script 
#errors
(1,36): expected-attribute-name-but-got-eof
(1,36): expected-named-closing-tag-but-got-eof
#new-errors
(1:37) eof-in-tag
#document
| <!DOCTYPE html>
| <script>
|   "<!--"

#data
<!doctype html><script><!--<s
#errors
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<s"

#data
<!doctype html><script><!--<script
#errors
(1,34): expected-named-closing-tag-but-got-eof
(1,34): unexpected-eof-in-text-mode
#new-errors
(1:35) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script"

#data
<!doctype html><script><!--<script 
#errors
(1,35): eof-in-script-in-script
(1,35): expected-named-closing-tag-but-got-eof
#new-errors
(1:36) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script "

#data
<!doctype html><script><!--<script <
#errors
(1,36): eof-in-script-in-script
(1,36): expected-named-closing-tag-but-got-eof
#new-errors
(1:37) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script <"

#data
<!doctype html><script><!--<script <a
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#new-errors
(1:38) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script <a"

#data
<!doctype html><script><!--<script </
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#new-errors
(1:38) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </"

#data
<!doctype html><script><!--<script </s
#errors
(1,38): eof-in-script-in-script
(1,38): expected-named-closing-tag-but-got-eof
#new-errors
(1:39) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </s"

#data
<!doctype html><script><!--<script </script
#errors
(1,43): eof-in-script-in-script
(1,43): expected-named-closing-tag-but-got-eof
#new-errors
(1:44) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </script"

#data
<!doctype html><script><!--<script </scripta
#errors
(1,44): eof-in-script-in-script
(1,44): expected-named-closing-tag-but-got-eof
#new-errors
(1:45) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </scripta"

#data
<!doctype html><script><!--<script </script 
#errors
(1,44): expected-named-closing-tag-but-got-eof
(1,44): unexpected-eof-in-text-mode
#new-errors
(1:45) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </script "

#data
<!doctype html><script><!--<script </script>
#errors
(1,44): expected-named-closing-tag-but-got-eof
(1,44): unexpected-eof-in-text-mode
#new-errors
(1:45) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </script>"

#data
<!doctype html><script><!--<script </script/
#errors
(1,44): expected-named-closing-tag-but-got-eof
(1,44): unexpected-eof-in-text-mode
#new-errors
(1:45) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </script/"

#data
<!doctype html><script><!--<script </script <
#errors
(1,45): expected-named-closing-tag-but-got-eof
(1,45): unexpected-eof-in-text-mode
#new-errors
(1:46) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </script <"

#data
<!doctype html><script><!--<script </script <a
#errors
(1,46): expected-named-closing-tag-but-got-eof
(1,46): unexpected-eof-in-text-mode
#new-errors
(1:47) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </script <a"

#data
<!doctype html><script><!--<script </script </
#errors
(1,46): expected-named-closing-tag-but-got-eof
(1,46): unexpected-eof-in-text-mode
#new-errors
(1:47) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </script </"

#data
<!doctype html><script><!--<script </script </script
#errors
(1,52): expected-named-closing-tag-but-got-eof
(1,52): unexpected-eof-in-text-mode
#new-errors
(1:53) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </script </script"

#data
<!doctype html><script><!--<script </script </script 
#errors
(1,53): expected-attribute-name-but-got-eof
(1,53): expected-named-closing-tag-but-got-eof
#new-errors
(1:54) eof-in-tag
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </script "

#data
<!doctype html><script><!--<script </script </script/
#errors
(1,53): unexpected-EOF-after-solidus-in-tag
(1,53): expected-named-closing-tag-but-got-eof
#new-errors
(1:54) eof-in-tag
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </script "

#data
<!doctype html><script><!--<script </script </script>
#errors
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script </script "

#data
<!doctype html><script><!--<script -
#errors
(1,36): eof-in-script-in-script
(1,36): expected-named-closing-tag-but-got-eof
#new-errors
(1:37) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script -"

#data
<!doctype html><script><!--<script -a
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#new-errors
(1:38) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script -a"

#data
<!doctype html><script><!--<script -<
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#new-errors
(1:38) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script -<"

#data
<!doctype html><script><!--<script --
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#new-errors
(1:38) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script --"

#data
<!doctype html><script><!--<script --a
#errors
(1,38): eof-in-script-in-script
(1,38): expected-named-closing-tag-but-got-eof
#new-errors
(1:39) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script --a"

#data
<!doctype html><script><!--<script --<
#errors
(1,38): eof-in-script-in-script
(1,38): expected-named-closing-tag-but-got-eof
#new-errors
(1:39) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script --<"

#data
<!doctype html><script><!--<script -->
#errors
(1,38): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script -->"

#data
<!doctype html><script><!--<script --><
#errors
(1,39): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script --><"

#data
<!doctype html><script><!--<script --></
#errors
(1,40): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script --></"

#data
<!doctype html><script><!--<script --></script
#errors
(1,46): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script --></script"

#data
<!doctype html><script><!--<script --></script 
#errors
(1,47): expected-attribute-name-but-got-eof
(1,47): expected-named-closing-tag-but-got-eof
#new-errors
(1:48) eof-in-tag
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script -->"

#data
<!doctype html><script><!--<script --></script/
#errors
(1,47): unexpected-EOF-after-solidus-in-tag
(1,47): expected-named-closing-tag-but-got-eof
#new-errors
(1:48) eof-in-tag
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script -->"

#data
<!doctype html><script><!--<script --></script>
#errors
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script -->"

#data
<!doctype html><script><!--<script><\/script>--></script>
#errors
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script><\/script>-->"

#data
<!doctype html><script><!--<script></scr'+'ipt>--></script>
#errors
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script></scr'+'ipt>-->"

#data
<!doctype html><script><!--<script></script><script></script></script>
#errors
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script></script><script></script>"

#data
<!doctype html><script><!--<script></script><script></script>--><!--</script>
#errors
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script></script><script></script>--><!--"

#data
<!doctype html><script><!--<script></script><script></script>-- ></script>
#errors
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script></script><script></script>-- >"

#data
<!doctype html><script><!--<script></script><script></script>- -></script>
#errors
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script></script><script></script>- ->"

#data
<!doctype html><script><!--<script></script><script></script>- - ></script>
#errors
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script></script><script></script>- - >"

#data
<!doctype html><script><!--<script></script><script></script>-></script>
#errors
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script></script><script></script>->"

#data
<!doctype html><script><!--<script>--!></script>X
#errors
(1,49): expected-named-closing-tag-but-got-eof
(1,49): unexpected-EOF-in-text-mode
#new-errors
(1:50) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script>--!></script>X"

#data
<!doctype html><script><!--<scr'+'ipt></script>--></script>
#errors
(1,59): unexpected-end-tag
#document
| <!DOCTYPE html>
| <script>
|   "<!--<scr'+'ipt>"
| "-->"

#data
<!doctype html><script><!--<script></scr'+'ipt></script>X
#errors
(1,57): expected-named-closing-tag-but-got-eof
(1,57): unexpected-eof-in-text-mode
#new-errors
(1:58) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <script>
|   "<!--<script></scr'+'ipt></script>X"

#data
<!doctype html><style><!--<style></style>--></style>
#errors
(1,52): unexpected-end-tag
#document
| <!DOCTYPE html>
| <style>
|   "<!--<style>"
| "-->"

#data
<!doctype html><style><!--</style>X
#errors
#document
| <!DOCTYPE html>
| <style>
|   "<!--"
| "X"

#data
<!doctype html><style><!--...</style>...--></style>
#errors
(1,51): unexpected-end-tag
#document
| <!DOCTYPE html>
| <style>
|   "<!--..."
| "...-->"

#data
<!doctype html><style><!--<br><html xmlns:v="urn:schemas-microsoft-com:vml"><!--[if !mso]><style></style>X
#errors
#document
| <!DOCTYPE html>
| <style>
|   "<!--<br><html xmlns:v="urn:schemas-microsoft-com:vml"><!--[if !mso]><style>"
| "X"

#data
<!doctype html><style><!--...<style><!--...--!></style>--></style>
#errors
(1,66): unexpected-end-tag
#document
| <!DOCTYPE html>
| <style>
|   "<!--...<style><!--...--!>"
| "-->"

#data
<!doctype html><style><!--...</style><!-- --><style>@import ...</style>
#errors
#document
| <!DOCTYPE html>
| <style>
|   "<!--..."
| <!--   -->
| <style>
|   "@import ..."

#data
<!doctype html><style>...<style><!--...</style><!-- --></style>
#errors
(1,63): unexpected-end-tag
#document
| <!DOCTYPE html>
| <style>
|   "...<style><!--..."
| <!--   -->

#data
<!doctype html><style>...<!--[if IE]><style>...</style>X
#errors
#document
| <!DOCTYPE html>
| <style>
|   "...<!--[if IE]><style>..."
| "X"

#data
<!doctype html><title><!--<title></title>--></title>
#errors
(1,52): unexpected-end-tag
#document
| <!DOCTYPE html>
| <title>
|   "<!--<title>"
| "-->"

#data
<!doctype html><title>&lt;/title></title>
#errors
#document
| <!DOCTYPE html>
| <title>
|   "</title>"

#data
<!doctype html><title>foo/title><link></head><body>X
#errors
(1,52): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <title>
|   "foo/title><link></head><body>X"

#data
<!doctype html><noscript><!--<noscript></noscript>--></noscript>
#errors
(1,64): unexpected-end-tag
#document
| <!DOCTYPE html>
| <noscript>
|   <!-- <noscript></noscript> -->

#data
<!doctype html><noscript><!--<noscript></noscript>--></noscript>
#errors
#document
| <!DOCTYPE html>
| <noscript>
|   <!-- <noscript></noscript> -->

#data
<!doctype html><noscript><!--</noscript>X<noscript>--></noscript>
#errors
#document
| <!DOCTYPE html>
| <noscript>
|   <!-- </noscript>X<noscript> -->

#data
<!doctype html><noscript><!--</noscript>X<noscript>--></noscript>
#errors
#document
| <!DOCTYPE html>
| <noscript>
|   <!-- </noscript>X<noscript> -->

#data
<!doctype html><noscript><iframe></noscript>X
#errors
#document
| <!DOCTYPE html>
| <noscript>
|   <iframe>
|     "</noscript>X"

#data
<!doctype html><noscript><iframe></noscript>X
#errors
 * (1,34) unexpected token in head noscript
 * (1,46) unexpected EOF
#document
| <!DOCTYPE html>
| <noscript>
|   <iframe>
|     "</noscript>X"

#data
<!doctype html><noframes><!--<noframes></noframes>--></noframes>
#errors
(1,64): unexpected-end-tag
#document
| <!DOCTYPE html>
| <noframes>
|   "<!--<noframes>"
| "-->"

#data
<!doctype html><noframes><body><script><!--...</script></body></noframes></html>
#errors
#document
| <!DOCTYPE html>
| <noframes>
|   "<body><script><!--...</script></body>"

#data
<!doctype html><textarea><!--<textarea></textarea>--></textarea>
#errors
(1,64): unexpected-end-tag
#document
| <!DOCTYPE html>
| <textarea>
|   "<!--<textarea>"
| "-->"

#data
<!doctype html><textarea>&lt;/textarea></textarea>
#errors
#document
| <!DOCTYPE html>
| <textarea>
|   "</textarea>"

#data
<!doctype html><textarea>&lt;</textarea>
#errors
#document
| <!DOCTYPE html>
| <textarea>
|   "<"

#data
<!doctype html><textarea>a&lt;b</textarea>
#errors
#document
| <!DOCTYPE html>
| <textarea>
|   "a<b"

#data
<!doctype html><iframe><!--<iframe></iframe>--></iframe>
#errors
(1,56): unexpected-end-tag
#document
| <!DOCTYPE html>
| <iframe>
|   "<!--<iframe>"
| "-->"

#data
<!doctype html><iframe>...<!--X->...<!--/X->...</iframe>
#errors
#document
| <!DOCTYPE html>
| <iframe>
|   "...<!--X->...<!--/X->..."

#data
<!doctype html><xmp><!--<xmp></xmp>--></xmp>
#errors
(1,44): unexpected-end-tag
#document
| <!DOCTYPE html>
| <xmp>
|   "<!--<xmp>"
| "-->"

#data
<!doctype html><noembed><!--<noembed></noembed>--></noembed>
#errors
(1,60): unexpected-end-tag
#document
| <!DOCTYPE html>
| <noembed>
|   "<!--<noembed>"
| "-->"

#data
<script>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,8): expected-named-closing-tag-but-got-eof
#document
| <script>

#data
<script>a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,9): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "a"

#data
<script><
#errors
(1,8): expected-doctype-but-got-start-tag
(1,9): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "<"

#data
<script></
#errors
(1,8): expected-doctype-but-got-start-tag
(1,10): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "</"

#data
<script></S
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "</S"

#data
<script></SC
#errors
(1,8): expected-doctype-but-got-start-tag
(1,12): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "</SC"

#data
<script></SCR
#errors
(1,8): expected-doctype-but-got-start-tag
(1,13): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "</SCR"

#data
<script></SCRI
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "</SCRI"

#data
<script></SCRIP
#errors
(1,8): expected-doctype-but-got-start-tag
(1,15): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "</SCRIP"

#data
<script></SCRIPT
#errors
(1,8): expected-doctype-but-got-start-tag
(1,16): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "</SCRIPT"

#data
<script></SCRIPT 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,17): expected-attribute-name-but-got-eof
(1,17): expected-named-closing-tag-but-got-eof
#new-errors
(1:18) eof-in-tag
#document
| <script>

#data
<script></s
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "</s"

#data
<script></sc
#errors
(1,8): expected-doctype-but-got-start-tag
(1,12): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "</sc"

#data
<script></scr
#errors
(1,8): expected-doctype-but-got-start-tag
(1,13): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "</scr"

#data
<script></scri
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "</scri"

#data
<script></scrip
#errors
(1,8): expected-doctype-but-got-start-tag
(1,15): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "</scrip"

#data
<script></script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,16): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "</script"

#data
<script></script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,17): expected-attribute-name-but-got-eof
(1,17): expected-named-closing-tag-but-got-eof
#new-errors
(1:18) eof-in-tag
#document
| <script>

#data
<script><!
#errors
(1,8): expected-doctype-but-got-start-tag
(1,10): expected-script-data-but-got-eof
(1,10): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "<!"

#data
<script><!a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "<!a"

#data
<script><!-
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "<!-"

#data
<script><!-a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,12): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "<!-a"

#data
<script><!--
#errors
(1,8): expected-doctype-but-got-start-tag
(1,12): expected-named-closing-tag-but-got-eof
(1,12): unexpected-eof-in-text-mode
#new-errors
(1:13) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--"

#data
<script><!--a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,13): expected-named-closing-tag-but-got-eof
(1,13): unexpected-eof-in-text-mode
#new-errors
(1:14) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--a"

#data
<script><!--<
#errors
(1,8): expected-doctype-but-got-start-tag
(1,13): expected-named-closing-tag-but-got-eof
(1,13): unexpected-eof-in-text-mode
#new-errors
(1:14) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<"

#data
<script><!--<a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
(1,14): unexpected-eof-in-text-mode
#new-errors
(1:15) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<a"

#data
<script><!--</
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
(1,14): unexpected-eof-in-text-mode
#new-errors
(1:15) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--</"

#data
<script><!--</script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,20): expected-named-closing-tag-but-got-eof
(1,20): unexpected-eof-in-text-mode
#new-errors
(1:21) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--</script"

#data
<script><!--</script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,21): expected-attribute-name-but-got-eof
(1,21): expected-named-closing-tag-but-got-eof
#new-errors
(1:22) eof-in-tag
#document
| <script>
|   "<!--"

#data
<script><!--<s
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
(1,14): unexpected-eof-in-text-mode
#new-errors
(1:15) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<s"

#data
<script><!--<script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,19): expected-named-closing-tag-but-got-eof
(1,19): unexpected-eof-in-text-mode
#new-errors
(1:20) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script"

#data
<script><!--<script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,20): eof-in-script-in-script
(1,20): expected-named-closing-tag-but-got-eof
#new-errors
(1:21) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script "

#data
<script><!--<script <
#errors
(1,8): expected-doctype-but-got-start-tag
(1,21): eof-in-script-in-script
(1,21): expected-named-closing-tag-but-got-eof
#new-errors
(1:22) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script <"

#data
<script><!--<script <a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,22): eof-in-script-in-script
(1,22): expected-named-closing-tag-but-got-eof
#new-errors
(1:23) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script <a"

#data
<script><!--<script </
#errors
(1,8): expected-doctype-but-got-start-tag
(1,22): eof-in-script-in-script
(1,22): expected-named-closing-tag-but-got-eof
#new-errors
(1:23) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script </"

#data
<script><!--<script </s
#errors
(1,8): expected-doctype-but-got-start-tag
(1,23): eof-in-script-in-script
(1,23): expected-named-closing-tag-but-got-eof
#new-errors
(1:24) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script </s"

#data
<script><!--<script </script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,28): eof-in-script-in-script
(1,28): expected-named-closing-tag-but-got-eof
#new-errors
(1:29) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script </script"

#data
<script><!--<script </scripta
#errors
(1,8): expected-doctype-but-got-start-tag
(1,29): eof-in-script-in-script
(1,29): expected-named-closing-tag-but-got-eof
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script </scripta"

#data
<script><!--<script </script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script </script "

#data
<script><!--<script </script>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script </script>"

#data
<script><!--<script </script/
#errors
(1,8): expected-doctype-but-got-start-tag
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script </script/"

#data
<script><!--<script </script <
#errors
(1,8): expected-doctype-but-got-start-tag
(1,30): expected-named-closing-tag-but-got-eof
(1,30): unexpected-eof-in-text-mode
#new-errors
(1:31) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script </script <"

#data
<script><!--<script </script <a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,31): expected-named-closing-tag-but-got-eof
(1,31): unexpected-eof-in-text-mode
#new-errors
(1:32) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script </script <a"

#data
<script><!--<script </script </
#errors
(1,8): expected-doctype-but-got-start-tag
(1,31): expected-named-closing-tag-but-got-eof
(1,31): unexpected-eof-in-text-mode
#new-errors
(1:32) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script </script </"

#data
<script><!--<script </script </script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,37): expected-named-closing-tag-but-got-eof
(1,37): unexpected-eof-in-text-mode
#new-errors
(1:38) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script </script </script"

#data
<script><!--<script </script </script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,38): expected-attribute-name-but-got-eof
(1,38): expected-named-closing-tag-but-got-eof
#new-errors
(1:39) eof-in-tag
#document
| <script>
|   "<!--<script </script "

#data
<script><!--<script </script </script/
#errors
(1,8): expected-doctype-but-got-start-tag
(1,38): unexpected-EOF-after-solidus-in-tag
(1,38): expected-named-closing-tag-but-got-eof
#new-errors
(1:39) eof-in-tag
#document
| <script>
|   "<!--<script </script "

#data
<script><!--<script </script </script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <script>
|   "<!--<script </script "

#data
<script><!--<script -
#errors
(1,8): expected-doctype-but-got-start-tag
(1,21): eof-in-script-in-script
(1,21): expected-named-closing-tag-but-got-eof
#new-errors
(1:22) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script -"

#data
<script><!--<script -a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,22): eof-in-script-in-script
(1,22): expected-named-closing-tag-but-got-eof
#new-errors
(1:23) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script -a"

#data
<script><!--<script --
#errors
(1,8): expected-doctype-but-got-start-tag
(1,22): eof-in-script-in-script
(1,22): expected-named-closing-tag-but-got-eof
#new-errors
(1:23) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script --"

#data
<script><!--<script --a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,23): eof-in-script-in-script
(1,23): expected-named-closing-tag-but-got-eof
#new-errors
(1:24) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script --a"

#data
<script><!--<script -->
#errors
(1,8): expected-doctype-but-got-start-tag
(1,23): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "<!--<script -->"

#data
<script><!--<script --><
#errors
(1,8): expected-doctype-but-got-start-tag
(1,24): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "<!--<script --><"

#data
<script><!--<script --></
#errors
(1,8): expected-doctype-but-got-start-tag
(1,25): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "<!--<script --></"

#data
<script><!--<script --></script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,31): expected-named-closing-tag-but-got-eof
#document
| <script>
|   "<!--<script --></script"

#data
<script><!--<script --></script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,32): expected-attribute-name-but-got-eof
(1,32): expected-named-closing-tag-but-got-eof
#new-errors
(1:33) eof-in-tag
#document
| <script>
|   "<!--<script -->"

#data
<script><!--<script --></script/
#errors
(1,8): expected-doctype-but-got-start-tag
(1,32): unexpected-EOF-after-solidus-in-tag
(1,32): expected-named-closing-tag-but-got-eof
#new-errors
(1:33) eof-in-tag
#document
| <script>
|   "<!--<script -->"

#data
<script><!--<script --></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <script>
|   "<!--<script -->"

#data
<script><!--<script><\/script>--></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <script>
|   "<!--<script><\/script>-->"

#data
<script><!--<script></scr'+'ipt>--></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <script>
|   "<!--<script></scr'+'ipt>-->"

#data
<script><!--<script></script><script></script></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <script>
|   "<!--<script></script><script></script>"

#data
<script><!--<script></script><script></script>--><!--</script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <script>
|   "<!--<script></script><script></script>--><!--"

#data
<script><!--<script></script><script></script>-- ></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <script>
|   "<!--<script></script><script></script>-- >"

#data
<script><!--<script></script><script></script>- -></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <script>
|   "<!--<script></script><script></script>- ->"

#data
<script><!--<script></script><script></script>- - ></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <script>
|   "<!--<script></script><script></script>- - >"

#data
<script><!--<script></script><script></script>-></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <script>
|   "<!--<script></script><script></script>->"

#data
<script><!--<script>--!></script>X
#errors
(1,8): expected-doctype-but-got-start-tag
(1,34): expected-named-closing-tag-but-got-eof
(1,34): unexpected-eof-in-text-mode
#new-errors
(1:35) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script>--!></script>X"

#data
<script><!--<scr'+'ipt></script>--></script>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,44): unexpected-end-tag
#document
| <script>
|   "<!--<scr'+'ipt>"
| "-->"

#data
<script><!--<script></scr'+'ipt></script>X
#errors
(1,8): expected-doctype-but-got-start-tag
(1,42): expected-named-closing-tag-but-got-eof
(1,42): unexpected-eof-in-text-mode
#new-errors
(1:43) eof-in-script-html-comment-like-text
#document
| <script>
|   "<!--<script></scr'+'ipt></script>X"

#data
<style><!--<style></style>--></style>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,37): unexpected-end-tag
#document
| <style>
|   "<!--<style>"
| "-->"

#data
<style><!--</style>X
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <style>
|   "<!--"
| "X"

#data
<style><!--...</style>...--></style>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,36): unexpected-end-tag
#document
| <style>
|   "<!--..."
| "...-->"

#data
<style><!--<br><html xmlns:v="urn:schemas-microsoft-com:vml"><!--[if !mso]><style></style>X
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <style>
|   "<!--<br><html xmlns:v="urn:schemas-microsoft-com:vml"><!--[if !mso]><style>"
| "X"

#data
<style><!--...<style><!--...--!></style>--></style>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,51): unexpected-end-tag
#document
| <style>
|   "<!--...<style><!--...--!>"
| "-->"

#data
<style><!--...</style><!-- --><style>@import ...</style>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <style>
|   "<!--..."
| <!--   -->
| <style>
|   "@import ..."

#data
<style>...<style><!--...</style><!-- --></style>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,48): unexpected-end-tag
#document
| <style>
|   "...<style><!--..."
| <!--   -->

#data
<style>...<!--[if IE]><style>...</style>X
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <style>
|   "...<!--[if IE]><style>..."
| "X"

#data
<title><!--<title></title>--></title>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,37): unexpected-end-tag
#document
| <title>
|   "<!--<title>"
| "-->"

#data
<title>&lt;/title></title>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <title>
|   "</title>"

#data
<title>foo/title><link></head><body>X
#errors
(1,7): expected-doctype-but-got-start-tag
(1,37): expected-named-closing-tag-but-got-eof
#document
| <title>
|   "foo/title><link></head><body>X"

#data
<noscript><!--<noscript></noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,49): unexpected-end-tag
#document
| <noscript>
|   <!-- <noscript></noscript> -->

#data
<noscript><!--<noscript></noscript>--></noscript>
#errors
 * (1,11) missing DOCTYPE
#document
| <noscript>
|   <!-- <noscript></noscript> -->

#data
<noscript><!--</noscript>X<noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <noscript>
|   <!-- </noscript>X<noscript> -->

#data
<noscript><!--</noscript>X<noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <noscript>
|   <!-- </noscript>X<noscript> -->

#data
<noscript><iframe></noscript>X
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <noscript>
|   <iframe>
|     "</noscript>X"

#data
<noscript><iframe></noscript>X
#errors
 * (1,11) missing DOCTYPE
 * (1,19) unexpected token in head noscript
 * (1,31) unexpected EOF
#document
| <noscript>
|   <iframe>
|     "</noscript>X"

#data
<noframes><!--<noframes></noframes>--></noframes>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,49): unexpected-end-tag
#document
| <noframes>
|   "<!--<noframes>"
| "-->"

#data
<noframes><body><script><!--...</script></body></noframes></html>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <noframes>
|   "<body><script><!--...</script></body>"

#data
<textarea><!--<textarea></textarea>--></textarea>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,49): unexpected-end-tag
#document
| <textarea>
|   "<!--<textarea>"
| "-->"

#data
<textarea>&lt;/textarea></textarea>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <textarea>
|   "</textarea>"

#data
<iframe><!--<iframe></iframe>--></iframe>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,41): unexpected-end-tag
#document
| <iframe>
|   "<!--<iframe>"
| "-->"

#data
<iframe>...<!--X->...<!--/X->...</iframe>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <iframe>
|   "...<!--X->...<!--/X->..."

#data
<xmp><!--<xmp></xmp>--></xmp>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,29): unexpected-end-tag
#document
| <xmp>
|   "<!--<xmp>"
| "-->"

#data
<noembed><!--<noembed></noembed>--></noembed>
#errors
(1,9): expected-doctype-but-got-start-tag
(1,45): unexpected-end-tag
#document
| <noembed>
|   "<!--<noembed>"
| "-->"

#data
<!doctype html><table>

#errors
(2,0): eof-in-table
#document
| <!DOCTYPE html>
| <table>
|   "
"

#data
<!doctype html><table><td><span><font></span><span>
#errors
(1,26): unexpected-cell-in-table-body
(1,45): unexpected-end-tag
(1,51): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <span>
|           <font>
|         <font>
|           <span>

#data
<!doctype html><form><table></form><form></table></form>
#errors
(1,35): unexpected-end-tag-implies-table-voodoo
(1,35): unexpected-end-tag
(1,41): unexpected-form-in-table
(1,56): unexpected-end-tag
(1,56): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <form>
|   <table>
|     <form>
